changed server/movie-service/internal/common/database/action/recommendation.go, server/movie-service/internal/lib/movie/data/movie.idx, server/movie-service/internal/lib/movie/idx.go, server/movie-service/internal/lib/movie/movie.go, server/movie-service/internal/types/types.go

+++ server/movie-service/internal/common/database/action/recommendation.go
@@ -10,7 +10,7 @@ import (
 	"gorm.io/gorm"
 )
 
-func CreateRecommendation(uuid string, genreID uint8) error {
+func CreateRecommendation(uuid string, genreID uint16) error {
 	logger := lib.NewLogger()
 	db := database.GetDB()
 
@@ -34,7 +34,7 @@ func CreateRecommendation(uuid string, genreID uint8) error {
 	return nil
 }
 
-func UpdateRecommendation(uuid string, genreID uint8) (bool, error) {
+func UpdateRecommendation(uuid string, genreID uint16) (bool, error) {
 	logger := lib.NewLogger()
 	db := database.GetDB()
