+++ server/movie-service/internal/types/types.go
@@ -1,7 +1,7 @@
 package types
 
 type RecommendationAddPayload struct {
-	Genres []uint8 `json:"genres" validate:"required"`
+	Genres []uint16 `json:"genres" validate:"required"`
 }
 
 type MoviePage struct {
@@ -16,18 +16,18 @@ type Movies struct {
 }
 
 type Movie struct {
-	Adult            bool    `json:"adult"`
-	BackdropPath     string  `json:"backdrop_path"`
-	GenreIds         []uint8 `json:"genre_ids"`
-	Id               uint32  `json:"id"`
-	OriginalLanguage string  `json:"original_language"`
-	OriginalTitle    string  `json:"original_title"`
-	Overview         string  `json:"overview"`
-	Popularity       float32 `json:"popularity"`
-	PosterPath       string  `json:"poster_path"`
-	ReleaseDate      string  `json:"release_date"`
-	Title            string  `json:"title"`
-	Video            bool    `json:"video"`
-	VoteAverage      float64 `json:"vote_average"`
-	VoteCount        uint16  `json:"vote_count"`
+	Adult            bool     `json:"adult"`
+	BackdropPath     string   `json:"backdrop_path"`
+	GenreIds         []uint16 `json:"genre_ids"`
+	Id               uint32   `json:"id"`
+	OriginalLanguage string   `json:"original_language"`
+	OriginalTitle    string   `json:"original_title"`
+	Overview         string   `json:"overview"`
+	Popularity       float32  `json:"popularity"`
+	PosterPath       string   `json:"poster_path"`
+	ReleaseDate      string   `json:"release_date"`
+	Title            string   `json:"title"`
+	Video            bool     `json:"video"`
+	VoteAverage      float64  `json:"vote_average"`
+	VoteCount        uint16   `json:"vote_count"`
 }
